Unifying statistic services
