--- src/main/java/com/vivacon/service/UserStatisticService.java
@@ -1,11 +1,0 @@
-package com.vivacon.service;
-
-import com.vivacon.dto.response.UserAccountMostFollower;
-
-import java.util.List;
-
-public interface UserStatisticService {
-
-    List<UserAccountMostFollower> getTheTopAccountMostFollowerStatistic(Integer limit);
-
-}
