+++ src/main/java/com/vivacon/service/impl/SettingServiceImpl.java
@@ -7,8 +7,13 @@ import com.vivacon.exception.RecordNotFoundException;
 import com.vivacon.repository.SettingRepository;
 import com.vivacon.service.AccountService;
 import com.vivacon.service.SettingService;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.NonTransientDataAccessException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.SQLException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -34,12 +39,16 @@ public class SettingServiceImpl implements SettingService {
     }
 
     @Override
+    @Transactional(propagation = Propagation.REQUIRED, rollbackFor = {DataIntegrityViolationException.class, NonTransientDataAccessException.class, SQLException.class, Exception.class})
     public boolean changeSetting(ChangeSettingRequest changeSettingRequest) {
         Long accountId = accountService.getCurrentAccount().getId();
         SettingType type = changeSettingRequest.getSettingType();
-        String value = type.serialize(changeSettingRequest.getValue());
-        return settingRepository.updateValueBySettingTypeAndAccountId(accountId, type,
-                changeSettingRequest.getValue()) > 0;
+        if (type.isValidValue(changeSettingRequest.getValue())) {
+            String value = type.serialize(changeSettingRequest.getValue());
+            return settingRepository.updateValueBySettingTypeAndAccountId(accountId, type, value) > 0;
+        } else {
+            throw new RuntimeException("Providing invalid value to change setting");
+        }
     }
 
     @Override
